+++ dashboard/src/router.js
@@ -2,11 +2,12 @@ import Vue from "vue";
 import Router from "vue-router";
 import Home from "./pages/Home";
 import Users from "./pages/Users";
+import Profile from "./pages/Profile";
+//import Content from "./pages/Content";
+//import Gallery from "./pages/Gallery";
+import Files from "./pages/Files";
 import SendEmail from "./pages/SendEmail";
 import GoogleMaps from "./pages/GoogleMaps";
-import Gallery from "./pages/Gallery";
-import Files from "./pages/Files";
-import Profile from "./pages/Profile";
 
 Vue.use(Router);
 
@@ -37,13 +38,13 @@ export default new Router({
         {
             path: "/gallery/:initPhotoSet",
             name: "gallery",
-            component: Gallery,
-            props: true
-
+            // component: Gallery,
             // route level code-splitting
             // this generates a separate chunk (about.[hash].js) for this route
             // which is lazy-loaded when the route is visited.
-            // component: () => import(/* webpackChunkName: "about" */ "../views/About.vue")
+            component: () =>
+                import(/* webpackChunkName: "gallery" */ "@/pages/Gallery.vue"),
+            props: true
         },
         {
             path: "/gallery",
@@ -53,6 +54,14 @@ export default new Router({
             path: "/files",
             name: "files",
             component: Files
+        },
+        {
+            path: "/content/:editContent?",
+            name: "content",
+            // component: Content,
+            component: () =>
+                import(/* webpackChunkName: "content" */ "@/pages/Content.vue"),
+            props: true
         },
         {
             path: "/profile/:initTab",