+++ dashboard/src/store/index.js
@@ -2,6 +2,7 @@ import Vue from "vue";
 import Vuex from "vuex";
 import users from "./modules/users";
 import profile from "./modules/profile";
+import content from "./modules/content";
 import gallery from "./modules/gallery";
 
 Vue.use(Vuex);
@@ -29,6 +30,7 @@ export default new Vuex.Store({
     modules: {
         users,
         profile,
+        content,
         gallery
     },
     state,