+++ facade/server/lib/api.js
@@ -4,26 +4,32 @@ const config = require("@config"),
     axios = require("axios").create({
         baseURL: config.api.baseUrl,
         headers: {
-            "Authorization": "Bearer " + config.api.token,
-            "X-Requested-With": "XMLHttpRequest"
-        }
+            Authorization: "Bearer " + config.api.token,
+            "X-Requested-With": "XMLHttpRequest",
+        },
     });
 
 axiosRetry(axios, {
     retries: 3,
     retryDelay: (retryCount, error) => {
-        logger.error("Api call error:", error.code, error.response ? error.response.status : "no response");
-        logger.info("Retry "+ retryCount);
+        logger.error(
+            "Api call error:",
+            error.code,
+            error.response ? error.response.status : "no response"
+        );
+        logger.info("Retry " + retryCount);
         if (!error) {
             return 0;
         }
-        if (error.code === "ECONNREFUSED"
-            || error.code === "ECONNRESET" 
-            || (error.response && error.response.status === 504)) {
+        if (
+            error.code === "ECONNREFUSED" ||
+            error.code === "ECONNRESET" ||
+            (error.response && error.response.status === 504)
+        ) {
             return retryCount * retryCount * 1000; // 1 sec, 4 sec, 9 sec
         }
         return 0;
-    }
+    },
 });
 
 class ApiError extends Error {
@@ -47,16 +53,33 @@ class ApiError extends Error {
 
 const call = new Proxy(axios, {
     get(target, name) {
-        const methods = ["request", "get", "delete", "head", "options", "post", "put", "patch"];
+        const methods = [
+            "request",
+            "get",
+            "delete",
+            "head",
+            "options",
+            "post",
+            "put",
+            "patch",
+        ];
         if (methods.includes(name) && typeof target[name] === "function") {
-            return async function() {
+            return async function () {
                 try {
                     let response = await target[name].apply(target, arguments);
-                    if (response && response.status >= 200 && response.status < 300) {
+                    if (
+                        response &&
+                        response.status >= 200 &&
+                        response.status < 300
+                    ) {
                         return response.data;
                     }
                 } catch (e) {
-                    if (e.response && e.response.status === 404 && e.response.statusText === e.response.data) {
+                    if (
+                        e.response &&
+                        e.response.status === 404 &&
+                        e.response.statusText === e.response.data
+                    ) {
                         return null;
                     }
                     throw new ApiError(e);
@@ -64,23 +87,29 @@ const call = new Proxy(axios, {
             };
         }
         return target[name];
-    }
+    },
 });
 
 module.exports.users = {
-    get: async usernameOrEmail => call.get("user/" + usernameOrEmail),
-    add: async user => call.post("user", user),
-    canRegister: async () => call.get("users/can-register")
+    get: async (usernameOrEmail) => call.get("user/" + usernameOrEmail),
+    add: async (user) => call.post("user", user),
+    canRegister: async () => call.get("users/can-register"),
+};
+
+module.exports.content = {
+    get: (code) => call.get("/content/" + code),
 };
 
 module.exports.photos = {
-    all: (sort, skip, limit) => call.get("/gallery/photos", { params: { sort, skip, limit } }),
+    all: (sort, skip, limit) =>
+        call.get("/gallery/photos", { params: { sort, skip, limit } }),
     count: () => call.get("/gallery/photos/count"),
-    get: id => call.get("/gallery/photo/" + id)
+    get: (id) => call.get("/gallery/photo/" + id),
 };
 
 module.exports.photoSets = {
     all: (sort) => call.get("/gallery/photosets", { params: { sort } }),
-    get: code => call.get("/gallery/photoset/" + code),
-    getWithPhotos: code => call.get(`/gallery/photoset/${code}`, { params: { photos: true } })
+    get: (code) => call.get("/gallery/photoset/" + code),
+    getWithPhotos: (code) =>
+        call.get(`/gallery/photoset/${code}`, { params: { photos: true } }),
 };